DOC: Update multimodal doc (#2785)
