+++ doc/source/locale/zh_CN/LC_MESSAGES/models/model_abilities/tools.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Xinference \n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-02-01 16:47+0800\n"
+"POT-Creation-Date: 2024-12-25 16:49+0800\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language: zh_CN\n"
@@ -17,7 +17,7 @@ msgstr ""
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel 2.13.1\n"
+"Generated-By: Babel 2.14.0\n"
 
 #: ../../source/models/model_abilities/tools.rst:5
 msgid "Tools"
@@ -42,8 +42,10 @@ msgid ""
 "define the functions along with their parameters and have the model "
 "dynamically choose which function to call and what parameters to pass to "
 "it."
-msgstr "就像 `OpenAI 的 Function calling API <https://platform.openai.com/docs/guides/function-calling>`_ 一样，"
-"你可以定义带有参数的函数，并让模型动态选择要调用哪个函数以及传递给它什么参数。"
+msgstr ""
+"就像 `OpenAI 的 Function calling API <https://platform.openai.com/docs/"
+"guides/function-calling>`_ 一样，你可以定义带有参数的函数，并让模型动态"
+"选择要调用哪个函数以及传递给它什么参数。"
 
 #: ../../source/models/model_abilities/tools.rst:19
 msgid "This is the general process for calling a function:"
@@ -62,15 +64,19 @@ msgid ""
 "inherent understanding or asking further details about the query and tool"
 " usage. On deciding to use a tool, it recommends the suitable API and "
 "instructions for its usage, framed in JSON."
-msgstr "LLM 决定是否启动功能。如果选择不启动，它会用日常语言回复，要么基于其内在理解提供解决方案，"
-"要么询问有关查询和工具使用的进一步细节。在决定使用工具时，它会推荐适合的 API 和 JSON 格式的使用说明。"
+msgstr ""
+"LLM 决定是否启动功能。如果选择不启动，它会用日常语言回复，要么基于其内在"
+"理解提供解决方案，要么询问有关查询和工具使用的进一步细节。在决定使用工具"
+"时，它会推荐适合的 API 和 JSON 格式的使用说明。"
 
 #: ../../source/models/model_abilities/tools.rst:25
 msgid ""
 "Following that, you implement the API call within your application and "
 "send the returned response back to the LLM for result analysis and "
 "proceeding with the next steps."
-msgstr "接下来，你在应用程序中实现 API 调用，并将返回的响应发送回 LLM 进行结果分析，并继续执行下一步操作。"
+msgstr ""
+"接下来，你在应用程序中实现 API 调用，并将返回的响应发送回 LLM 进行结果"
+"分析，并继续执行下一步操作。"
 
 #: ../../source/models/model_abilities/tools.rst:28
 msgid ""
@@ -93,54 +99,68 @@ msgid ":ref:`models_llm_qwen-chat`"
 msgstr ""
 
 #: ../../source/models/model_abilities/tools.rst:36
-msgid ":ref:`models_llm_chatglm3`"
+msgid ":ref:`models_llm_glm4-chat`"
 msgstr ""
 
 #: ../../source/models/model_abilities/tools.rst:37
-msgid ":ref:`models_llm_gorilla-openfunctions-v1`"
+msgid ":ref:`models_llm_glm4-chat-1m`"
 msgstr ""
 
-#: ../../source/models/model_abilities/tools.rst:41
+#: ../../source/models/model_abilities/tools.rst:38
+msgid ":ref:`models_llm_llama-3.1-instruct`"
+msgstr ""
+
+#: ../../source/models/model_abilities/tools.rst:39
+msgid ":ref:`models_llm_qwen2.5-instruct`"
+msgstr ""
+
+#: ../../source/models/model_abilities/tools.rst:40
+msgid ":ref:`models_llm_qwen2.5-coder-instruct`"
+msgstr ""
+
+#: ../../source/models/model_abilities/tools.rst:43
 msgid "Quickstart"
 msgstr "快速入门"
 
-#: ../../source/models/model_abilities/tools.rst:43
+#: ../../source/models/model_abilities/tools.rst:45
 msgid ""
 "An optional parameter ``tools`` in the Chat API can be used to provide "
 "function specifications. The purpose of this is to enable models to "
 "generate function arguments which adhere to the provided specifications."
-msgstr "Chat API 中的可选参数 ``tools`` 可以用于提供函数规范。其目的是使模型能够生成符合所提供规范的函数参数。"
+msgstr ""
+"Chat API 中的可选参数 ``tools`` 可以用于提供函数规范。其目的是使模型能够"
+"生成符合所提供规范的函数参数。"
 
-#: ../../source/models/model_abilities/tools.rst:47
+#: ../../source/models/model_abilities/tools.rst:49
 msgid "Example using OpenAI Client"
 msgstr "使用 OpenAI 客户端的示例"
 
-#: ../../source/models/model_abilities/tools.rst:96
+#: ../../source/models/model_abilities/tools.rst:98
 msgid "The output will be:"
 msgstr "输出结果是："
 
-#: ../../source/models/model_abilities/tools.rst:115
+#: ../../source/models/model_abilities/tools.rst:117
 msgid ""
 "Finish reason will be ``tool_calls`` if the LLM uses a tool call. "
 "Othewise it will be the default finish reason."
-msgstr "如果 LLM 使用了工具调用，完成原因将是 ``tool_calls`` 。"
-"否则，它将是默认的完成原因。"
+msgstr ""
+"如果 LLM 使用了工具调用，完成原因将是 ``tool_calls`` 。否则，它将是默认的"
+"完成原因。"
 
-#: ../../source/models/model_abilities/tools.rst:120
+#: ../../source/models/model_abilities/tools.rst:122
 msgid ""
 "The API will not actually execute any function calls. It is up to "
 "developers to execute function calls using model outputs."
 msgstr "API 本身不会执行任何函数调用。开发者需要使用模型输出来执行函数调用。"
 
-#: ../../source/models/model_abilities/tools.rst:124
+#: ../../source/models/model_abilities/tools.rst:126
 msgid "You can find more examples of ``tools`` ability in the tutorial notebook:"
 msgstr "你可以在教程笔记本中找到更多关于 ``tools`` 能力的示例。"
 
-#: ../../source/models/model_abilities/tools.rst:128
+#: ../../source/models/model_abilities/tools.rst:130
 msgid "Function calling"
 msgstr "函数调用"
 
-#: ../../source/models/model_abilities/tools.rst:131
+#: ../../source/models/model_abilities/tools.rst:133
 msgid "Learn from a complete example demonstrating function calling"
 msgstr "学习一个完整的示例，演示函数调用的过程。"
-
