--- doc/source/locale/zh_CN/LC_MESSAGES/models/model_abilities/vision.po
@@ -1,126 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) 2023, Xorbits Inc.
-# This file is distributed under the same license as the Xinference package.
-# FIRST AUTHOR <EMAIL@ADDRESS>, 2024.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: Xinference \n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-08-20 11:56+0800\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language: zh_CN\n"
-"Language-Team: zh_CN <LL@li.org>\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel 2.14.0\n"
-
-#: ../../source/models/model_abilities/vision.rst:5
-msgid "Vision"
-msgstr "视觉"
-
-#: ../../source/models/model_abilities/vision.rst:7
-msgid "Learn how to process images with LLMs."
-msgstr "学习如何使用 LLM 处理图像。"
-
-#: ../../source/models/model_abilities/vision.rst:11
-msgid "Introduction"
-msgstr "介绍"
-
-#: ../../source/models/model_abilities/vision.rst:13
-msgid ""
-"With the ``vision`` ability you can have your model take in images and "
-"answer questions about them. Within Xinference, this indicates that "
-"certain models are capable of processing image inputs when conducting "
-"dialogues via the Chat API."
-msgstr ""
-"通过 ``vision`` 能力，您可以让模型接收图像并回答有关它们的问题。在 "
-"Xinference 中，这表示某些模型在通过 Chat API 进行对话时能够处理图像输入。"
-
-#: ../../source/models/model_abilities/vision.rst:19
-msgid "Supported models"
-msgstr "支持的模型列表"
-
-#: ../../source/models/model_abilities/vision.rst:21
-msgid ""
-"The ``vision`` ability is supported with the following models in "
-"Xinference:"
-msgstr "在 Xinference 中支持 ``vision`` 功能的模型如下："
-
-#: ../../source/models/model_abilities/vision.rst:23
-msgid ":ref:`qwen-vl-chat <models_llm_qwen-vl-chat>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:24
-msgid ":ref:`deepseek-vl-chat <models_llm_deepseek-vl-chat>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:25
-msgid ":ref:`yi-vl-chat <models_llm_yi-vl-chat>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:26
-msgid ":ref:`omnilmm <models_llm_omnilmm>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:27
-msgid ":ref:`internvl-chat <models_llm_internvl-chat>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:28
-msgid ":ref:`cogvlm2 <models_llm_cogvlm2>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:29
-msgid ":ref:`MiniCPM-Llama3-V 2.5 <models_llm_minicpm-llama3-v-2_5>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:30
-msgid ":ref:`GLM-4V <models_llm_glm-4v>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:31
-msgid ":ref:`MiniCPM-Llama3-V 2.6 <models_llm_minicpm-v-2.6>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:32
-msgid ":ref:`internvl2 <models_llm_internvl2>`"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:36
-msgid "Quickstart"
-msgstr "快速入门"
-
-#: ../../source/models/model_abilities/vision.rst:38
-msgid ""
-"Images are made available to the model in two main ways: by passing a "
-"link to the image or by passing the base64 encoded image directly in the "
-"request."
-msgstr ""
-"模型可以通过两种主要方式获取图像：通过传递图像的链接或直接在请求中传递 "
-"base64 编码的图像。"
-
-#: ../../source/models/model_abilities/vision.rst:42
-msgid "Example using OpenAI Client"
-msgstr "使用 OpenAI 客户端的示例"
-
-#: ../../source/models/model_abilities/vision.rst:73
-msgid "Uploading base 64 encoded images"
-msgstr "上传 Base64 编码的图片"
-
-#: ../../source/models/model_abilities/vision.rst:115
-msgid "You can find more examples of ``vision`` ability in the tutorial notebook:"
-msgstr "你可以在教程笔记本中找到更多关于 ``vision`` 能力的示例。"
-
-#: ../../source/models/model_abilities/vision.rst:119
-msgid "Qwen VL Chat"
-msgstr ""
-
-#: ../../source/models/model_abilities/vision.rst:122
-msgid "Learn vision ability from a example using qwen-vl-chat"
-msgstr "通过使用 qwen-vl-chat 的示例来学习使用 LLM 的视觉能力"
-
