+++ doc/source/locale/zh_CN/LC_MESSAGES/user_guide/backends.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Xinference \n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-08-20 11:56+0800\n"
+"POT-Creation-Date: 2025-01-26 11:51+0800\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language: zh_CN\n"
@@ -140,8 +140,9 @@ msgstr "目前，支持的模型包括："
 
 #: ../../source/user_guide/backends.rst:49
 msgid ""
-"``llama-2``, ``llama-3``, ``llama-3.1``, ``llama-2-chat``, "
-"``llama-3-instruct``, ``llama-3.1-instruct``"
+"``llama-2``, ``llama-3``, ``llama-3.1``, ``llama-3.2-vision``, "
+"``llama-2-chat``, ``llama-3-instruct``, ``llama-3.1-instruct``, "
+"``llama-3.3-instruct``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:50
@@ -166,70 +167,93 @@ msgstr ""
 #: ../../source/user_guide/backends.rst:54
 msgid ""
 "``deepseek``, ``deepseek-coder``, ``deepseek-chat``, ``deepseek-coder-"
-"instruct``"
+"instruct``, ``deepseek-r1-distill-qwen``, ``deepseek-v2-chat``, "
+"``deepseek-v2-chat-0628``, ``deepseek-v2.5``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:55
-msgid "``codeqwen1.5``, ``codeqwen1.5-chat``"
+msgid "``yi-coder``, ``yi-coder-chat``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:56
-msgid "``baichuan-2-chat``"
+msgid "``codeqwen1.5``, ``codeqwen1.5-chat``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:57
-msgid "``internlm2-chat``"
+msgid ""
+"``qwen2.5``, ``qwen2.5-coder``, ``qwen2.5-instruct``, ``qwen2.5-coder-"
+"instruct``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:58
-msgid "``internlm2.5-chat``, ``internlm2.5-chat-1m``"
+msgid "``baichuan-2-chat``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:59
-msgid "``qwen-chat``"
+msgid "``internlm2-chat``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:60
-msgid "``mixtral-instruct-v0.1``, ``mixtral-8x22B-instruct-v0.1``"
+msgid "``internlm2.5-chat``, ``internlm2.5-chat-1m``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:61
-msgid "``chatglm3``, ``chatglm3-32k``, ``chatglm3-128k``"
+msgid "``qwen-chat``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:62
-msgid "``glm4-chat``, ``glm4-chat-1m``"
+msgid "``mixtral-instruct-v0.1``, ``mixtral-8x22B-instruct-v0.1``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:63
-msgid "``codegeex4``"
+msgid "``chatglm3``, ``chatglm3-32k``, ``chatglm3-128k``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:64
-msgid "``qwen1.5-chat``, ``qwen1.5-moe-chat``"
+msgid "``glm4-chat``, ``glm4-chat-1m``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:65
-msgid "``qwen2-instruct``, ``qwen2-moe-instruct``"
+msgid "``codegeex4``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:66
-msgid "``gemma-it``, ``gemma-2-it``"
+msgid "``qwen1.5-chat``, ``qwen1.5-moe-chat``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:67
-msgid "``orion-chat``, ``orion-chat-rag``"
+msgid "``qwen2-instruct``, ``qwen2-moe-instruct``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:68
-msgid "``c4ai-command-r-v01``"
+msgid "``QwQ-32B-Preview``"
+msgstr ""
+
+#: ../../source/user_guide/backends.rst:69
+msgid "``marco-o1``"
+msgstr ""
+
+#: ../../source/user_guide/backends.rst:70
+msgid "``gemma-it``, ``gemma-2-it``"
+msgstr ""
+
+#: ../../source/user_guide/backends.rst:71
+msgid "``orion-chat``, ``orion-chat-rag``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:72
-msgid "SGLang"
+msgid "``c4ai-command-r-v01``"
 msgstr ""
 
 #: ../../source/user_guide/backends.rst:73
+msgid "``minicpm3-4b``"
+msgstr ""
+
+#: ../../source/user_guide/backends.rst:77
+msgid "SGLang"
+msgstr ""
+
+#: ../../source/user_guide/backends.rst:78
 msgid ""
 "`SGLang <https://github.com/sgl-project/sglang>`_ has a high-performance "
 "inference runtime with RadixAttention. It significantly accelerates the "
@@ -241,11 +265,11 @@ msgstr ""
 " 的高性能推理运行时。它通过在多个调用之间自动重用KV缓存，显著加速了复杂 "
 "LLM 程序的执行。它还支持其他常见推理技术，如连续批处理和张量并行处理。"
 
-#: ../../source/user_guide/backends.rst:78
+#: ../../source/user_guide/backends.rst:83
 msgid "MLX"
 msgstr ""
 
-#: ../../source/user_guide/backends.rst:79
+#: ../../source/user_guide/backends.rst:84
 msgid ""
 "`MLX <https://github.com/ml-explore/mlx-examples/tree/main/llms>`_ "
 "provides efficient runtime to run LLM on Apple silicon. It's recommended "
@@ -256,29 +280,4 @@ msgstr ""
 "苹果 silicon 芯片上高效运行 LLM 的方式。在模型包含 MLX 格式的时候，推荐"
 "使用苹果 silicon 芯片的 Mac 用户使用 MLX 引擎。"
 
-#~ msgid "``llama-2``, ``llama-3``, ``llama-2-chat``, ``llama-3-instruct``"
-#~ msgstr ""
-
-#~ msgid "``mistral-v0.1``, ``mistral-instruct-v0.1``, ``mistral-instruct-v0.2``"
-#~ msgstr ""
-
-#~ msgid ""
-#~ "``llama-2``, ``llama-3``, ``llama-2-chat``, "
-#~ "``llama-3-instruct``, ``llama-3.1``, "
-#~ "``llama-3.1-instruct``"
-#~ msgstr ""
-
-#~ msgid "``gemma-it``"
-#~ msgstr ""
-
-#~ msgid "``baichuan``, ``baichuan-chat``, ``baichuan-2-chat``"
-#~ msgstr ""
-
-#~ msgid ""
-#~ "``internlm-16k``, ``internlm-chat-7b``, "
-#~ "``internlm-chat-8k``, ``internlm-chat-20b``"
-#~ msgstr ""
-
-#~ msgid "``vicuna-v1.3``, ``vicuna-v1.5``"
-#~ msgstr ""
