--- doc/source/models/builtin/audio/fishspeech-1.4.rst
@@ -1,19 +1,0 @@
-.. _models_builtin_fishspeech-1.4:
-
-==============
-FishSpeech-1.4
-==============
-
-- **Model Name:** FishSpeech-1.4
-- **Model Family:** FishAudio
-- **Abilities:** text-to-audio
-- **Multilingual:** True
-
-Specifications
-^^^^^^^^^^^^^^
-
-- **Model ID:** fishaudio/fish-speech-1.4
-
-Execute the following command to launch the model::
-
-   xinference launch --model-name FishSpeech-1.4 --model-type audio
