+++ doc/source/models/builtin/llm/index.rst
@@ -437,7 +437,7 @@ The following is a list of built-in LLM in Xinference:
      - Qwen1.5-MoE is a transformer-based MoE decoder-only language model pretrained on a large amount of data.
 
    * - :ref:`qwen2-audio <models_llm_qwen2-audio>`
-     - chat, audio
+     - generate, audio
      - 32768
      - Qwen2-Audio: A large-scale audio-language model which is capable of accepting various audio signal inputs and performing audio analysis or direct textual responses with regard to speech instructions.
